Reword pod priority and preemption concept (#17508)
* Remove mentions of unsupported Kubernetes versions

No need to mention supported-since-this-version details for Kubernetes
releases that are now out of support.

* Warn about disabling preemption

* Tweak wording

* Add What's next section

* Tweak Pod priority troubleshooting advice

Reshape the advice about pod priority troubleshooting to explain
user-induced issues (and define expected behavior). If the reader
detects behavior that does not match the documentation, they have
observed a bug and can report that via the usual routes.

* Rewords notes on Pod priority vs. QoS

* Move warning into page body