Update setup-ha-etcd-with-kubeadm.md
It was not clear to me that I had to add the unit file to all nodes and restart the kubelet service. I was stuck wondering why etcd could not connect to the other peers until i realized that kubelet was only running on `HOST-0`.

Remove parantheses.

Improve phrasing.

Update content/en/docs/setup/production-environment/tools/kubeadm/setup-ha-etcd-with-kubeadm.md

Co-authored-by: Tim Bannister <tim@scalefactory.com>