Caution Note for ssh-auth Secrets
Adds a `caution` note that SSH key pairs do not establish trust between
clients and servers. A secondary method is required to establish trust
between an SSH client and host server, such as fixed `known_hosts` file.
Clients which do not establish adequate trust are vulnerable to "man in
the middle" impersonation attacks.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>