remove misleading instructions
There is no need to restart the kubelet as part of configuring kubeadm to set the cgroup driver. At this point in the setup instructions, the kubernetes cluster isn't even up and running yet, as kubeadm init hasn't been run.  The previous step even says the kubelet is in a crashloop waiting for kubeadm.