Reformat windows introduction
The current markdown text for the Windows introduction page is very
difficult for tracking changes. There are many paragraphs shown as a
single line. We need line breaks to ease the change tracking, especially
for localization teams. There are other nits related to indentation and
numbered lists. This PR is an attempt to reformat the page so that
future revisions can use it as an example.