Fix: Update service.md (nginx image not found)
I fixed an imagepulling error in the service.md example on the website. nginx:11.14.2 is no longer on Nginx docker hub:  https://hub.docker.com/_/nginx?tab=tags&page=1&name=11.14.2 so I changed the image tag to the latest. please let me know if the latest tag is not the right image to use or if I'm wrong.