Improve page-meta-links page
This PR fixes a warning in the page-meta-links.html page where `.Path`
is a function to be removed soon. Warning message:

```
WARN 2022/06/17 21:20:35 .Path when the page is backed by a file is deprecated and will be removed in a future release. We plan to use Path for a canonical source path and you probably want to check the source is a file.
```