Fix quantity case for ephemeral storage
All resource requests parse units the same way. Ephemeral storage previously had an upper case `K` as a valid unit but the kubernetes server can not parse this value.

Similar to memory, add a statement regarding that a lowercase `m` will be accepted by the kubernetes server, but is probably an incorrect request.