Fix in-cluster API discovery documentation (#36691)
* Fix in-cluster API discovery documentation

The documentation incorrectly describes the way that client libraries
discover the Kubernetes API server. While the `kubernetes.default.svc`
DNS is provided as a convenience, **all** of the officially supported API
clients use environment variables to discover the address of the API server.

This change updates the documentation to reflect this.

Signed-off-by: Oliver Gould <ver@buoyant.io>

* Review feedback

* Fixup

Signed-off-by: Oliver Gould <ver@buoyant.io>