Fix generated config API for API server encryption
There are limitations in the generator when parsing ambiguous strings
such as `*`, `<` etc. `*` is used in Markdown for emphasis, but some
comments are using the literal '*' without properly escaping it.
`<` and `>` are not allowed in Markdown parser. The generated content
for this is at best some warning strings.

This PR fixes this issue manually.

An ideal fix would be proposed to the upstream source code wherever
these happen.