Add reference for kubectl command
The prvious [kubectl reference](https://kubernetes.io/docs/reference/generated/kubectl/kubectl-commands)
has some problems to resolve and we had never had the luck invest
resource into it. The issues include:

- The kubectl reference is a HTML file, with its own look & feel that
  is different from the rest of the website.
- The kubectl generator is using a container `brianpursley/brodocs:latest`
  which has not seen any maintenance for over severl years. ... and now,
  it is broken. The `make cli` command no longer works, and we don't
  know where the source code is.
- The current kubectl reference is not suitable for browsing on small
  devices. A huge portion of the screen asset is occupied by the
  examples layer. We don't like it and we can do nothing about it.

The added reference for kubectl is generated using the same generator
we use for other k8s components, so the source code is in the
'kubernetes-sig/reference-docs" projects. There is no additional effort
to generate them. The only thing we need to do for each release is to
update it just like we do for other k8s components.