Update NVIDIA GPU scheduling link
The link to the NVIDIA GPU vendor implementation reference for the
scheduling-gpus page was pointing at their gpu-feature-discovery repo.
This repo is deprecated and the work has moved to the k8s-device-plugin
repo.

Luckily the old repo has a migration notice that leads folks to the new
location, so they ultimately are able to get the information they are
looking for. This updates the link to take out the extra step of visting
the old repo first.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>